New code to read CSV files of names/tokens
Separated from the fitbit_authenticate since the OAuth piece is a
separate one-time process. Did not actually change the authentication
file until this code is working. This is not quite there yet - the loop
is not reading for all lines in the file, but it can read an
access_token and read a name to write the results file.